Modulated searchbar tests into their own file
